Added a model class for Job and an interface to expose services.
Created a Job class in models with multiple attributes and made this class as public so that, this class can be accessed in other classes especially in services.
Job class contains three attributes id, name and description with setters and getters to read and manipulate data.

Created a IJobDataStore interface under services to expose services to read or write job data.
This interface has methods for the following CRUD operations.
1.	Created GetJobs() to fetch all the jobs.
2.	Created GetJob() by passing id to fetch a particular job.
3.	Created AddJob() to insert new job to the database
4.	Created UpdateJob() to update a specific job in the table.