01:41:35  Implemented methods in JobListViewModel and created a JobListPage
Added methods in JobListViewModel class to fetch and display all jobs, refresh jobs, select a particular job.
Modified the default constructor to set title of the page, load jobs and set refresh and select commands.
Implemented methods for selecting a job and refresh jobs.
Refresh command will reload the jobs.
Select command will redirect JobListPage  with details of the selected job in a private object.
Created a LoadJobs method to fetch all the jobs from the JobDataStore.
Created a new content page under views and named it as JobListPage to display the list of jobs.
Updated JobListPage.xaml to bind the JobListViewModel to the view.
Injected the refresh command to the RefreshView.
Created a scrollable list view with basic styling to display the list of jobs.