02:19:20 Implemented CRUD operations and minor fixes.
Commented selected method code in JobListViewModel and modified LoadJobs method.
Enabled Xaml Compilation in JobListPage.
Added list view item template to display the name of each job in the job list page.
Added file-based storage to the application. Created a JSON file to read and write job data.
Created methods to read and write data to the JSON file using serialize and deserialize.
Implemented below functionalities:
AddJob which reads the existing data, appends the new job and write the updated data to JSON.
DeleteJob which reads the existing data, finds the particular job, based on jobId and remove that job from file.
GetJob which reads the existing data, finds the particular job based on jobId.
GetJobs which reads the existing data, return all jobs in the file
UpdateJob which reads the existing data, finds the particular job based on jobId and write the updated data to JSON.