Added Xamarin dependencies and created new context pages.
Added Xamarin dependencies for Forms, Community Toolkit and Community Markup.

Added a new context page i.e., WelcomePage to display welcome message in the app. Created WelcomeViewModel class and binded in WelcomePage.xaml.

Created a new ViewModel for JobDetails and named it as JobDetailViewModel. The view model is created with one attribute jobId and corresponding getter and setter. Created binding to this view model in JobDetailPage.xaml. Added a label to the content template to display the Job Id.

Modified AppShell.xaml to add FlyoutItem for Welcome page and JobDetail page with corresponding route and content template details.