Added azure storage blobs dependency to integrate with Azure storage account for Job data storage.
Created a resource in azure and accessed key and connection string from azure storage account to use in the application

Created a new implementation (JobDataStoreBlobStorageJson) of IJobDataStore interface to integrate with azure storage. Added the connection string acquired from storage account, defined the container name and file name.

Created a service client of type BlobServiceClient by passing the connection string to interact with azure storage services.

Implemented WriteFile method to write the list of jobs to the  azure store using factory methods from BlobServiceClient.

Added the default methods for the rest of the methods of the interface.